Moving target enums to a separate file
